Route Page: Spinner AND Autocomplete Text
Showing you guys an example of how autocomplete would work. right now I
just need to set it to do something when the field is selected. (it pops
up the list automatically right now when the app is opened, without that
line of code, it will show a list when input is not there (aka threshold
is 0).