Working Database!
Very exciting. The editFavorites screen now passes info to the database
when you press the save button, and when you go to FavoritesActivity
(which was used only for testing; will be changed to a popup later) it
shows the favorites you just made! So cool. Thanks to Dave for the help!