redesigned UI
needs to be refined, but now the app starts on the LocateInput screen,
the menu button brings up edit favorites, maps, and about. for some
reason the back button doesn't see a menu press as something to go back
to, so I have implented a quickndirty menu button to get back to the
previous screen. I need to do more research on why the back button is
behaving like this.